+++ src/cli/cli_help.c
@@ -1,0 +1,5 @@
+void
+cli_help()
+{
+	return;
+}
